Certificate part, creation of keys
For this commit, we tried to create (from different types) in order to have the desire setup.

the creation for this certificates were done by openssl, aditionally it was planned to create a wildcard SSL for this behavior, neverthless as errors occurred, it was not possible.

Hence, I finnaly created 3 certificates (each for domain name) and moreover installed crt,key into the server and .crt to the client.

Command: openssl req -new -x509 -keyout server.key -out server.crt -days 365 -nodes