Client summary
IThe client.py was created by an image on Docker, this image contains all the information of the aplication and also the certificates that server contains.

Additionally, this client was supossed to connect via DNS queries by default, and furthermore I tried to take more steps and have a connectivity usign DNS resolution by Kubernetes (Services and namespaces).

For the last point, I had use host aliasses in order to follow the stream path that takes to resolve the 3 queries. I couldnt get the authentication for the client